feat: consolidate email operations into EmailController (#299)
Move EmailOutbox, Pause/Resume, Retry/Discard, and EmailPreview from
AdminController to new EmailController at /Email. Facilitated messaging
(HumanController.SendMessage, CampController.Contact) remains contextual
by design. AdminController is now significantly smaller.
