+++ src/Humans.Web/Views/Email/EmailPreview.cshtml
@@ -11,7 +11,7 @@
 
 <nav aria-label="breadcrumb">
     <ol class="breadcrumb">
-        <li class="breadcrumb-item"><a asp-action="Index">Admin</a></li>
+        <li class="breadcrumb-item"><a asp-controller="Admin" asp-action="Index">Admin</a></li>
         <li class="breadcrumb-item active" aria-current="page">Email Preview</li>
     </ol>
 </nav>
